ai_db/bin/requirements.txt
# requirements.txt
# Legal AI 통합 시스템 의존성

# FastAPI
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3

# AI/ML
torch==2.1.2
transformers==4.36.2
sentence-transformers==2.2.2
faiss-cpu==1.7.4

# Data
pandas==2.1.4
pyarrow==14.0.2

# LLM
google-generativeai==0.3.2

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.25

# Utils
python-dotenv==1.0.0
